services/web/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
gunicorn
flask-cors